refactoring

+++ MyServer/ClientHandler.cpp
@@ -224,7 +224,7 @@ unsigned long ClientHandler::myrecv(int socket, char **data) {
             for(int i = 0; i < sizeReceived; i++){
                 byteBuffer.push_back(buffer[i]);
             }
-            if(byteBuffer.size() < size){
+            if((unsigned)byteBuffer.size() < size){
                 send(socket, "\0", BUF, 0); //send confirmation
             }
         }
